backend/services/ai-guidance/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
httpx==0.25.0
redis==5.0.1
pydantic==2.4.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncio==3.4.3
asyncio-mqtt==0.16.1
websockets==11.0.3
tenacity==8.2.3
aiofiles==23.2.1
Pillow==10.0.1
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.1
numpy==1.25.2
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.12.2
plotly==5.17.0
google-generativeai==0.3.1
playwright==1.39.0
beautifulsoup4==4.12.2
requests==2.31.0
jinja2==3.1.2
python-dotenv==1.0.0
pdfplumber==0.9.0
PyJWT==2.8.0
google-cloud-aiplatform==1.38.0
